extract common part for symbol search for scala 2 and 3
